+++ utils/build.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+SCRIPT_ROOT=$(cd $(dirname $0); pwd)
+cd $SCRIPT_ROOT/..
+
+npx tsc -v && npx tsc -b \
+    packages/library \
+    packages/tests/* \
+    packages/examples/*