+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes Logs
 
+## v0.3.0
+
+- chore: updated to TypeScript v6.
+- chore: updated all dependencies to latest version.
+- build: switched to monorepo structure.
+- refactor: migrate to ESM module system.
+
 ## v0.2.2
 
 - fix(typings): fixed definition of `IGeneralDecorator` adapting to TS5.0.